- Comment out SLACK_CHANGE which sent STUN binding (badly formatted) to keep interfaces alive. That was an attempt to get some mobility and did not work well. Also, Chrome started logging bad STUN packets because of that - Introduce a new constant SLACK_CHANGE_NO_CONNCHECK to avoid connection check for discovered peer reflexive candidate. This might have been the issue with clients getting confused between relay candidate and peer reflexive candidate and shutting down the relay candidate. Clients create a peer reflexive candidate when they see an address that they do not recognize. This is a bit of a guess looking at logs and client code. But, it is still puzzling.
Here is what happens (for sendonly connections)
- Client announces a fake candidate
- Server announces good candidate which is an internal address (
reachable only through TURN server)
- Client talks to TURN server and gets a TURN allocation/binding
(there is a STUN binding which is actuallly handled by the TURN server
also)
- There are further STUN bindings from the client which are forwarded
by the TURN server to the media server.
- libnice looks at those and add a new remote address (note that libnice
is seeing a first proper candidate from the client as previously it was
announcing fake candidate in the SDP) as a discovered peer reflexive
addresses
- libnice replies to the STUN binding request from the client
- libnice then send STUN binding request to the new discovered address
in strict accordance with ICE standard to ensure that the reverse path works.
- Client could see a new address and add a peer reflexive candidate

Puzzling bits
- Does client add a candidate/connection when it hears back from the TURN
server and what is that address?
- Or does it add the TURN candidate/connection to its list when it hears
back STUN binding response from libnice
- If the latter, it is possible that STUN binding request from libnice comes
earlier (either due to out of order packet or lost packet). So, by cutting
off the STUN binding request from libnice side, we are eliminating this path.
- For some  people, it happens like clockwork, why?